Tries to upgrade the customimzed html elements
We are overwriting some default elements to display everything more
coherently. These overwritten files were not updated to the newer
version, as they are packaged in the theme gem package. This commit
upgrades them to their newer version, porting any of our changes.

Two changes to the existing custom changes are:
1. hiding the search bar instead of removing it so that the pwa script
   does not break
2. add line breaks before imprint/privacy policy