Overwrite search-loader.html to suppress search.json warning
The search-loader.html file includes a jekyll search plugin, which
tries to load a search.json file. Since we are not providing the file,
it always reported an error in the dev consolve. By overwriting
search-loader.html with an empty file, the search engine does not get
included. This supresses the warning and breaks the search, but since it
is not included anyway, this is not a problem.